Move the mir_toolkit/event enums to libmircore-dev and reduce miral header dependencies to that.
