+++ examples/example-server-lib/floating_window_manager.cpp
@@ -24,6 +24,8 @@
 #include <miral/window_info.h>
 #include <miral/window_manager_tools.h>
 
+#include <mir_toolkit/event.h> // @arg TODO
+
 #include <linux/input.h>
 #include <csignal>
