+++ examples/example-server-lib/floating_window_manager.h
@@ -23,7 +23,7 @@
 
 #include "splash_session.h"
 
-#include <mir_toolkit/event.h> // @arg TODO
+#include <mir_toolkit/events/enums.h>
 
 #include <chrono>
 #include <map>
