+++ examples/miral-kiosk/kiosk_window_manager.h
@@ -23,7 +23,7 @@
 
 #include <miral/canonical_window_manager.h>
 
-#include <mir_toolkit/event.h> // @arg TODO
+#include <mir_toolkit/events/enums.h>
 
 using namespace mir::geometry;
